api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.11
chromadb==0.4.24
sentence-transformers==2.5.1
pypdf==4.1.0
python-dotenv==1.0.0
gitpython==3.1.41